requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.4.0
Django==1.11.20
django-cors-headers==3.0.2
django-filter==2.0.0
djangorestframework==3.9.2
enum34==1.1.6
ipdb==0.12
ipython==5.8.0
ipython-genutils==0.2.0
Markdown==3.1.1
nose==1.3.7
pathlib2==2.3.3
pexpect==4.7.0
php==1.2.0
phply==1.2.5
pickleshare==0.7.5
ply==3.11
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2019.1
scandir==1.10.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
zipfile2==0.0.12
